Add Validators, Refactoring, Bump to .net 7

+++ Directory/DirectoryCopier.cs
@@ -1,10 +1,13 @@
 ﻿using System;
+using System.Collections.Generic;
 using System.IO;
 using System.Security;
 using System.Text;
+using Humanizer;
 using PhotoCopy.Files;
+using PhotoCopy.Validators;
 
-namespace PhotoCopySort
+namespace PhotoCopy.Directory
 {
     /// <summary>
     /// Handles copying of files in directory.
@@ -15,21 +18,23 @@ namespace PhotoCopySort
         /// Copies files from input folder to output folder based on options.
         /// </summary>
         /// <param name="options">Options</param>
-        public static void Copy(Options options)
+        public static void Copy(Options options, IReadOnlyCollection<IValidator> validators)
         {
             foreach (var file in DirectoryScanner.EnumerateFiles(options.Source, options))
             {
                 Log.Print($">> {file.File.FullName}", Options.LogLevel.verbose);
 
+                if (ShouldCopy(validators, file))
+                {
+                    continue;
+                }
+
                 var newPath = GeneratePath(options, file);
                 var newFile = new FileInfo(newPath);
 
-                if (newFile.Exists)
+                if (!ResolveDuplicate(options, file, ref newFile))
                 {
-                    if (!ResolveDuplicate(options, file, ref newFile))
-                    {
-                        continue;
-                    }
+                    continue;
                 }
 
                 var directory = newFile.Directory;
@@ -51,18 +56,37 @@ namespace PhotoCopySort
             }
         }
 
+        private static bool ShouldCopy(IReadOnlyCollection<IValidator> validators, IFile file)
+        {
+            foreach (var validator in validators)
+            {
+                if (!validator.Validate(file))
+                {
+                    Log.Print($"\tFiltered by {validator.GetType().Name.Humanize().ToLowerInvariant()}", Options.LogLevel.verbose);
+                    return true;
+                }
+            }
+
+            return false;
+        }
+
         private static bool ResolveDuplicate(Options options, IFile file, ref FileInfo newFile)
         {
+            if (!newFile.Exists)
+            {
+                return true;
+            }
+
             if (options.SkipExisting)
             {
                 return false;
             }
 
             var newGenericFile = new GenericFile(newFile,
-                new FileDateTime {DateTime = newFile.CreationTime});
+                new FileDateTime { DateTime = newFile.CreationTime });
             if (!options.NoDuplicateSkip && EqualChecksum(file, newGenericFile))
             {
-                Log.Print($"Duplicate of {newFile.FullName}", Options.LogLevel.verbose);
+                Log.Print($"\tDuplicate of {newFile.FullName}", Options.LogLevel.verbose);
                 if (!options.DryRun && options.Mode == Options.OperationMode.move)
                 {
                     file.File.Delete();
@@ -105,15 +129,15 @@ namespace PhotoCopySort
         private static string GeneratePath(Options options, IFile source)
         {
             var builder = new StringBuilder(options.Destination)
-                .Replace(Options.DestinationEnum.Year, source.FileDateTime.DateTime.Year.ToString())
-                .Replace(Options.DestinationEnum.Month, source.FileDateTime.DateTime.Month.ToString())
-                .Replace(Options.DestinationEnum.Day, source.FileDateTime.DateTime.Day.ToString())
-                .Replace(Options.DestinationEnum.DayOfYear, source.FileDateTime.DateTime.DayOfYear.ToString())
-                .Replace(Options.DestinationEnum.Directory,
+                .Replace(Options.DestinationVariables.Year, source.FileDateTime.DateTime.Year.ToString())
+                .Replace(Options.DestinationVariables.Month, source.FileDateTime.DateTime.Month.ToString())
+                .Replace(Options.DestinationVariables.Day, source.FileDateTime.DateTime.Day.ToString())
+                .Replace(Options.DestinationVariables.DayOfYear, source.FileDateTime.DateTime.DayOfYear.ToString())
+                .Replace(Options.DestinationVariables.Directory,
                     Path.GetRelativePath(options.Source, source.File.DirectoryName ?? ""))
-                .Replace(Options.DestinationEnum.Name, source.File.Name)
-                .Replace(Options.DestinationEnum.NameNoExtension, Path.GetFileNameWithoutExtension(source.File.Name))
-                .Replace(Options.DestinationEnum.Extension, source.File.Extension.TrimStart('.'));
+                .Replace(Options.DestinationVariables.Name, source.File.Name)
+                .Replace(Options.DestinationVariables.NameNoExtension, Path.GetFileNameWithoutExtension(source.File.Name))
+                .Replace(Options.DestinationVariables.Extension, source.File.Extension.TrimStart('.'));
 
 
             return builder.ToString();
